block: move struct request to blk-mq.h
struct request is only used by blk-mq drivers, so move it and all
related declarations to blk-mq.h.

Link: https://lore.kernel.org/r/20210920123328.1399408-18-hch@lst.de
